chore: move branded icons to images
