+++ src/dappsExplorer/DappCard.tsx
@@ -1,14 +1,14 @@
 import React from 'react'
 import { Image, StyleSheet, Text, View, ViewStyle } from 'react-native'
-import { DappExplorerEvents } from 'src/analytics/Events'
 import AppAnalytics from 'src/analytics/AppAnalytics'
+import { DappExplorerEvents } from 'src/analytics/Events'
 import Card from 'src/components/Card'
 import Touchable from 'src/components/Touchable'
 import { favoriteDappIdsSelector } from 'src/dapps/selectors'
 import { favoriteDapp, unfavoriteDapp } from 'src/dapps/slice'
 import { Dapp, DappSection } from 'src/dapps/types'
-import Star from 'src/icons/Star'
 import StarOutline from 'src/icons/StarOutline'
+import Star from 'src/images/Star'
 import { useDispatch, useSelector } from 'src/redux/hooks'
 import { Colors } from 'src/styles/colors'
 import { typeScale } from 'src/styles/fonts'
