+++ src/earn/EarnInfoScreen.tsx
@@ -9,11 +9,11 @@ import Button, { BtnSizes, BtnTypes } from 'src/components/Button'
 import { APP_NAME, EARN_STABLECOINS_LEARN_MORE } from 'src/config'
 import { EarnTabType } from 'src/earn/types'
 import ArrowDown from 'src/icons/ArrowDown'
-import Blob from 'src/icons/Blob'
 import CircledIcon from 'src/icons/CircledIcon'
 import EarnCoins from 'src/icons/EarnCoins'
-import Palm from 'src/icons/Palm'
+import Blob from 'src/images/Blob'
 import Logo from 'src/images/Logo'
+import Palm from 'src/images/Palm'
 import { headerWithCloseButton } from 'src/navigator/Headers'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
