+++ src/home/NotificationBox.tsx
@@ -17,8 +17,8 @@ import { dismissNotification } from 'src/home/actions'
 import { DEFAULT_PRIORITY } from 'src/home/reducers'
 import { getExtraNotifications } from 'src/home/selectors'
 import { Notification, NotificationBannerCTATypes, NotificationType } from 'src/home/types'
-import GuideKeyIcon from 'src/icons/GuideKeyHomeCardIcon'
 import KeylessBackup from 'src/icons/KeylessBackup'
+import GuideKeyIcon from 'src/images/GuideKeyHomeCardIcon'
 import { getVerified, learnCelo } from 'src/images/Images'
 import { ensurePincode, navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
