+++ src/images/NftReceivedIcon.tsx
