+++ src/images/Palm.tsx
