+++ src/images/Star.tsx
