+++ src/images/WaveCurve.tsx
