+++ src/jumpstart/JumpstartAddAssets.tsx
@@ -8,8 +8,8 @@ import { BottomSheetRefType } from 'src/components/BottomSheet'
 import Button, { BtnSizes, BtnTypes } from 'src/components/Button'
 import CustomHeader from 'src/components/header/CustomHeader'
 import { CICOFlow, FiatExchangeFlow } from 'src/fiatExchanges/utils'
-import Palm from 'src/icons/Palm'
-import WaveCurve from 'src/icons/WaveCurve'
+import Palm from 'src/images/Palm'
+import WaveCurve from 'src/images/WaveCurve'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { isAppSwapsEnabledSelector } from 'src/navigator/selectors'
