+++ src/onboarding/registration/ProtectWallet.tsx
@@ -6,10 +6,10 @@ import { SafeAreaView } from 'react-native-safe-area-context'
 import { recoveryPhraseInOnboardingStarted } from 'src/account/actions'
 import { RecoveryPhraseInOnboardingStatus } from 'src/account/reducer'
 import { recoveryPhraseInOnboardingStatusSelector } from 'src/account/selectors'
-import { OnboardingEvents } from 'src/analytics/Events'
 import AppAnalytics from 'src/analytics/AppAnalytics'
+import { OnboardingEvents } from 'src/analytics/Events'
 import OnboardingCard from 'src/components/OnboardingCard'
-import GuideKeyIcon from 'src/icons/GuideKeyIcon'
+import GuideKeyIcon from 'src/images/GuideKeyIcon'
 import { HeaderTitleWithSubtitle, nuxNavigationOptionsOnboarding } from 'src/navigator/Headers'
 import { ensurePincode, navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
