+++ src/transactions/feed/NftFeedItem.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { StyleSheet, Text, View } from 'react-native'
-import { HomeEvents } from 'src/analytics/Events'
 import AppAnalytics from 'src/analytics/AppAnalytics'
+import { HomeEvents } from 'src/analytics/Events'
 import IconWithNetworkBadge from 'src/components/IconWithNetworkBadge'
 import Touchable from 'src/components/Touchable'
 import ImageErrorIcon from 'src/icons/ImageErrorIcon'
-import NftReceivedIcon from 'src/icons/NftReceivedIcon'
-import NftSentIcon from 'src/icons/NftSentIcon'
+import NftReceivedIcon from 'src/images/NftReceivedIcon'
+import NftSentIcon from 'src/images/NftSentIcon'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import NftMedia from 'src/nfts/NftMedia'
